graspoutine

+++ CONTROLLERS/controller_admin.php
@@ -44,8 +44,9 @@ class controller_admin extends controller_main
         {
             throw new Exception('Pseudo vide');
         }
-        echo $userSupp->getState();
-        exit();
+        if  ($userSupp != null && $userSupp->getState() == 'admin') {
+            throw new Exception('Admin impossible a supprimer');
+        }
         $this->user->deleteUser();
         $vue = new Vue('userDeleted');
         $vue->generer(array('param'=> $userSupp));
